chore: improving application use cases and header based authentication

+++ src/main/java/com/labs/paycore/wallet/application/WalletDepositUseCaseInput.java
@@ -1,9 +1,8 @@
 package com.labs.paycore.wallet.application;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 public record WalletDepositUseCaseInput(
-  BigDecimal amount,
+  long amount,
   UUID userId
 ) {}
