+++ src/main/java/com/labs/paycore/wallet/application/WalletP2PUseCaseInput.java
@@ -1,10 +1,9 @@
 package com.labs.paycore.wallet.application;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 public record WalletP2PUseCaseInput(
   UUID senderWalletId,
-  BigDecimal amount,
+  long amount,
   String recipient
 ) {}
